Joining existing network

+++ include/InfoMessage.h
@@ -1,13 +1,14 @@
 #ifndef SRC_INFOMESSAGE_H_
 #define SRC_INFOMESSAGE_H_
+#include <cstdlib>
 
 class InfoMessage {
 public:
-	InfoMessage(unsigned opcode = 0);
-	unsigned opcode;
-	unsigned firstField;
-	unsigned secondField;
-	unsigned thirdField;
+	InfoMessage(size_t opcode = 0);
+	size_t opcode;
+	size_t firstField;
+	size_t secondField;
+	size_t thirdField;
 };
 
 #endif /* SRC_INFOMESSAGE_H_ */
